First file that gets executed when initiated.
This is the main File which gets executed when the ESP32 is started for the first time. It takes Various Inputs such as USERNAME,SSID,SSID_PASSWORD and appliance Names from the User and store it in a txt file. We have kept One Reset Interrupt. It gets executed once we press the Boot Pin of the ESP32 for 5 consecutive seconds, the contents of the data file gets deleted. And we need to re-run the boot file in order to make this system work.